refactor code, update README
