+++ src/hooks/useFetchResources.js
@@ -9,12 +9,14 @@ import {
   getChartConfig,
   getElmDependencies,
   getInterventionLogicLib,
+  isNumber,
+} from "../util/util";
+import {
   getResourcesByResourceType,
   getFhirResourcesFromQueryResult,
   getFHIRResourcesToLoad,
   getFHIRResourcePaths,
-  isNumber,
-} from "../util/util";
+} from "../util/fhirUtil";
 import qConfig from "../config/questionnaire_config";
 
 export default function useFetchResources() {
